skillcompass_backend/requirements.txt
python-jose[cryptography]
passlib[bcrypt]
pytest
httpx
slowapi
fastapi==0.115.12
uvicorn==0.34.2
firebase-admin==6.8.0
python-dotenv==1.0.1
openai

